Updated project architecture and compiler + English
The compiler now knows how to deal with translations and has better
debugging capabilities (it prints wrong lines). The first prototype was
placed in the src folder.

The program is only runnable with the included compiler. The source
files rely on the syntax features the advanced mode (-a) provides.

Also integrated the CONST and the TITR program into the source and
started to update the integrated strings to the new language-agnostic
system.